requirements.txt
PyQRCode~=1.2.1
Pillow~=10.1.0
qrcode~=7.4.2